+++ src/main/java/com/jss/osiris/modules/osiris/quotation/service/CentralPayDelegateService.java
@@ -1,5 +1,7 @@
 package com.jss.osiris.modules.osiris.quotation.service;
 
+import java.math.BigDecimal;
+
 import com.jss.osiris.libs.exception.OsirisException;
 import com.jss.osiris.modules.osiris.quotation.model.centralPay.CentralPayPaymentRequest;
 import com.jss.osiris.modules.osiris.quotation.model.centralPay.CentralPayTransaction;
@@ -8,7 +10,7 @@ public interface CentralPayDelegateService {
 
         public CentralPayPaymentRequest cancelPaymentRequest(String centralPayPaymentRequestId);
 
-        public CentralPayPaymentRequest generatePayPaymentRequest(Double amount, String mail, String entityId,
+        public CentralPayPaymentRequest generatePayPaymentRequest(BigDecimal amount, String mail, String entityId,
                         String subject);
 
         public CentralPayPaymentRequest getPaymentRequest(String centralPayPaymentRequestId) throws OsirisException;
